+++ units/host_monitoring_container_consistency_test.go
@@ -2,28 +2,26 @@ package units
 
 import (
 	"context"
+	"path/filepath"
 	"testing"
 
 	"github.com/evergreen-ci/evergreen"
 	"github.com/evergreen-ci/evergreen/db"
-	"github.com/evergreen-ci/evergreen/mock"
 	"github.com/evergreen-ci/evergreen/model/host"
 	"github.com/evergreen-ci/evergreen/testutil"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
 
 func TestHostMonitoringContainerConsistencyJob(t *testing.T) {
 	assert := assert.New(t)
-	require := require.New(t)
 	testConfig := testutil.TestConfig()
 	db.SetGlobalSessionProvider(testConfig.SessionFactory())
 
 	assert.NoError(db.Clear(host.Collection))
 
-	env := &mock.Environment{
-		EvergreenSettings: testConfig,
-	}
+	ctx := context.Background()
+	env := evergreen.GetEnvironment()
+	assert.NoError(env.Configure(ctx, filepath.Join(evergreen.FindEvergreenHome(), testutil.TestDir, testutil.TestSettings), nil))
 
 	h1 := &host.Host{
 		Id:            "parent-1",
@@ -38,13 +36,13 @@ func TestHostMonitoringContainerConsistencyJob(t *testing.T) {
 	h3 := &host.Host{
 		Id:       "container-2",
 		Status:   evergreen.HostRunning,
-		ParentID: "parent-2",
+		ParentID: "parent-1",
 	}
 	assert.NoError(h1.Insert())
 	assert.NoError(h2.Insert())
 	assert.NoError(h3.Insert())
 
-	j := NewHostMonitorExternalStateJob(env, h1, evergreen.ProviderNameDockerMock)
+	j := NewHostMonitorContainerStateJob(env, h1, evergreen.ProviderNameDockerMock, "job-1")
 	assert.False(j.Status().Completed)
 
 	j.Run(context.Background())
@@ -54,9 +52,9 @@ func TestHostMonitoringContainerConsistencyJob(t *testing.T) {
 
 	container1, err := host.FindOne(host.ById("container-1"))
 	assert.NoError(err)
-	assert.Equal(container1.Status, evergreen.HostRunning)
+	assert.Equal(evergreen.HostRunning, container1.Status)
 
 	container2, err := host.FindOne(host.ById("container-2"))
 	assert.NoError(err)
-	assert.Equal(container2.Status, evergreen.HostTerminated)
+	assert.Equal(evergreen.HostTerminated, container2.Status)
 }
